Move ViewCellUtils to util package
