+++ viewcelladapter/src/main/java/ca/antonious/viewcelladapter/utils/ViewCellUtils.java
@@ -1,4 +1,4 @@
-package ca.antonious.viewcelladapter;
+package ca.antonious.viewcelladapter.utils;
 
 import java.util.List;
